More improvements to connection lifetime handling * Fixed issue where I was polling the client socket before it was connected * When a client socket connection fails for whatever reason we now clean up the connection info properly and stop attempting to send data to the client * Added notion of "connection state" to the server request handler. This gives each connection a place to store persistent state. Right now this is only used to store which controller streams a given connection id has active. This structure will most likely move out of the request handler and into a more central location.

+++ psmovedataframe/PackedMessage.h
@@ -111,7 +111,7 @@ public:
             return false;
 
         int msg_size = m_msg->ByteSize();
-        if (HEADER_SIZE + msg_size < buf_size)
+        if ((int)HEADER_SIZE + msg_size < buf_size)
         {
             memset(buf, 0, buf_size);
             encode_header(buf, buf_size, msg_size);
